Using unsigned short for undo identifiers as they can't collide with NSNotFound.

+++ App/SVTextDOMController.m
@@ -317,7 +317,7 @@
 
 @implementation SVWebEditorTextControllerUndoManager
 
-- (NSUInteger)lastRegisteredActionIdentifier;
+- (unsigned short)lastRegisteredActionIdentifier;
 {
     return _lastRegisteredActionIdentifier;
 }
